the data sending doesnt work, but the API def works

+++ models/purchase.go
@@ -5,7 +5,7 @@ import "gorm.io/gorm"
 type Purchase struct {
 	gorm.Model
 	Name     string  `json:"name"`
-	Cost     float64 `json:"cost"`
+	Cost     float32 `json:"cost"`
 	Category string  `json:"category"`
 	UserID   uint    `json:"user_id"`
 }
